app/requirements.txt
# --- PyTorch (CUDA 12.4) ---
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchvision==0.21.0
transformers>=4.45.0,<5
accelerate>=0.29.0
huggingface_hub>=0.24.0
timm>=0.9.0
streamlit==1.38.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.26.0
supervision>=0.18.0


# --- SAM ---
segment-anything==1.0

# --- Optionnel ---
sentencepiece>=0.1.99
bitsandbytes==0.43.1
huggingface_hub>=0.24.0
